requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
tensorflow-gpu==2.2.0
mtcnn
imutils
Pillow
opencv-python
numpy